requirements.txt
grpcio==1.66.2
grpcio-tools==1.66.2
psycopg2-binary==2.9.9